# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Radio call manager API — real-time audio streaming, call management, and admin CRUD.",
        "title": "OpenScanner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/admin/export/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the entire server configuration as a JSON download. Sensitive fields (API key hashes, downstream API keys, webhook secrets) are excluded.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Export full configuration",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.configExport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/import/config": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Imports settings, groups, tags, systems, talkgroups, units, API keys, dirwatches, downstreams, and webhooks from a JSON body. Existing records are upserted or skipped on conflict.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Import full configuration",
                "parameters": [
                    {
                        "description": "Configuration data to import",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.configImport"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok: true",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/setup": {
            "post": {
                "description": "Creates the initial admin user and marks setup as complete. Returns 409 if setup is already done, 400 for invalid input.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup"
                ],
                "summary": "Complete initial setup",
                "parameters": [
                    {
                        "description": "Admin credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.setupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "ok": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/setup/status": {
            "get": {
                "description": "Returns whether initial setup is needed and whether public access is active. Always unauthenticated.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup"
                ],
                "summary": "Get setup status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.setupStatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_openscanner_openscanner_internal_db.Dirwatch": {
            "type": "object",
            "properties": {
                "delay": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "delete_after": {
                    "type": "integer"
                },
                "directory": {
                    "type": "string"
                },
                "disabled": {
                    "type": "integer"
                },
                "extension": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "frequency": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "id": {
                    "type": "integer"
                },
                "mask": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "order": {
                    "type": "integer"
                },
                "system_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "talkgroup_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "type": {
                    "type": "string"
                },
                "use_polling": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_openscanner_openscanner_internal_db.Group": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "github_com_openscanner_openscanner_internal_db.Setting": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_openscanner_openscanner_internal_db.System": {
            "type": "object",
            "properties": {
                "auto_populate": {
                    "type": "integer"
                },
                "blacklists_json": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "led": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "order": {
                    "type": "integer"
                },
                "system_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_openscanner_openscanner_internal_db.Tag": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "github_com_openscanner_openscanner_internal_db.Talkgroup": {
            "type": "object",
            "properties": {
                "frequency": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "group_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "led": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "name": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "order": {
                    "type": "integer"
                },
                "system_id": {
                    "type": "integer"
                },
                "tag_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "talkgroup_id": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "descriptive error message"
                }
            }
        },
...
        "internal_api.configExport": {
            "type": "object",
            "properties": {
                "apiKeys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.exportAPIKey"
                    }
                },
                "dirwatches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Dirwatch"
                    }
                },
                "downstreams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.exportDownstream"
                    }
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Group"
                    }
                },
                "settings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Setting"
                    }
                },
                "systems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.System"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Tag"
                    }
                },
                "talkgroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Talkgroup"
                    }
                },
                "units": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Unit"
                    }
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.User"
                    }
                },
                "webhooks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.exportWebhook"
                    }
                }
            }
        },
        "internal_api.configImport": {
            "type": "object",
            "properties": {
                "apiKeys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.ApiKey"
                    }
                },
                "dirwatches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Dirwatch"
                    }
                },
                "downstreams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Downstream"
                    }
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Group"
                    }
                },
                "settings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Setting"
                    }
                },
                "systems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.System"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Tag"
                    }
                },
                "talkgroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Talkgroup"
                    }
                },
                "units": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Unit"
                    }
                },
                "webhooks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_openscanner_openscanner_internal_db.Webhook"
                    }
                }
            }
        },
...
        "internal_api.exportAPIKey": {
            "type": "object",
            "properties": {
                "disabled": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ident": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "systemsJson": {
                    "type": "string"
                }
            }
        },
        "internal_api.exportDownstream": {
            "type": "object",
            "properties": {
                "disabled": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "systemsJson": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.setupRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_api.setupStatusResponse": {
            "type": "object",
            "properties": {
                "needsSetup": {
                    "type": "boolean"
                },
                "publicAccess": {
                    "type": "boolean"
                }
            }
        },
...
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "API key for call upload endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token prefixed with \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}